add test for test_vertexai_multimodal_embedding_text_input
